requirements.txt
# Web Framework and API
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-multipart==0.0.9  # For form data handling

# Database
pymongo==4.6.1
motor==3.3.2  # Async MongoDB driver

# Google APIs
google-api-python-client==2.118.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-generativeai==0.3.2

# Email Processing
beautifulsoup4==4.12.3

# HTTP and Environment
requests==2.31.0
python-dotenv==1.0.1
httpx==0.27.0  # For async HTTP requests

# Testing
pytest==8.0.2
pytest-asyncio==0.23.5  # For async tests
pytest-cov==4.1.0  # For test coverage

# Development
black==24.2.0  # Code formatting
isort==5.13.2  # Import sorting
flake8==7.0.0  # Linting

# Logging
loguru==0.7.2

# New added pydantic-settings
pydantic-settings==2.1.0